Try to make a binary_search that is always producing the lower bound
I didn't intend for this to be the final version, only the first working
version of a binary_search that always returns the lower bound. But it
seems to perform well (please test and see if it performs exactly as
well on your machine).

Code adapted from
http://en.cppreference.com/w/cpp/algorithm/lower_bound; the first
example, so not so much thinking went into it.